Plot landscape

This function plots a specific time step landscape

plot_landscape(land, nRow, nCol, Title = NULL, rmBorder = TRUE,
  rangeLimit = NULL)

Arguments

land
a vector with the distribution of the four forest states in the landscape. This is found in the output list from the run_model function named 'land_TX', where X is the time step

nRow
numeric, number of rows of the landscape. Value is found in the output list from the run_model function

nCol
numeric, number of columns of the landscape. Value is found in the output list from the run_model function

Title
character, title of the landscape plot

rmBorder
logical, if TRUE the four side borders will be removed of the plot. This option is available because the model do not calculate state prevalence in the borders.

rangeLimit,
vector, the latitudinal position of the boreal and temperate range limit in the landscape configuration. This value is obtained from the function range_limit

Examples

# NOT RUN {
plot_landscape(land = output[[100]], nRow = output[['nRow']],
               nCol = output[['nCol']], Title = 'land at step 100')
# }