Calculate state range limit

This function calculates the south range limit of boreal and the north range limit of temperate states in a specific landscape configuration

range_limit(land, nRow, nCol, occup)

Arguments

land
vector, one element of the run_model output

nRow
numeric, number of rows of the landscape. Value is found in the output list from the run_model function

nCol
numeric, number of columns of the landscape. Value is found in the output list from the run_model function

occup
numeric between 0 and 1. The value determines the minimum occupancy a row of the landscape must be occupied by a specific forest state to be considered part of the state range

Examples

if (FALSE) {
lands <- run_model(steps = 10, initLand)
range_limit(lands[['land_TX']], lands[['nRow']], lands[['nCol']], occup = .7)
}